+++ src/main/java/ru/mipt/bit/platformer/util/Direction.java
@@ -1,0 +1,21 @@
+package ru.mipt.bit.platformer.util;
+
+import com.badlogic.gdx.math.GridPoint2;
+
+public class Direction {
+    private final GridPoint2 deltaCoordinate;
+    private float angle;
+
+    public Direction(int x, int y) {
+        deltaCoordinate = new GridPoint2(x, y);
+        angle = (float) (Math.atan2(y, x) * 180 / Math.PI);
+    }
+
+    public float getAngle() {
+        return angle;
+    }
+
+    public GridPoint2 getDeltaCoordinate() {
+        return new GridPoint2(deltaCoordinate);
+    }
+}
